# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CoopDigital — API de gestion des coopératives cacaoyères
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /api/refus/count:
    get:
      operationId: countRefusEnAttente
      tags: [refus]
      summary: Nombre de refus en attente de traitement
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Comptage
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
  /api/refus/stats:
    get:
      operationId: getStatsRefus
      tags: [refus]
      summary: Statistiques stocks refoulés
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefusStats"
  /api/refus:
    get:
      operationId: listRefus
      tags: [refus]
      summary: Liste des stocks refoulés
      security: [{ bearerAuth: [] }]
      parameters:
        - name: statut
          in: query
          schema:
            type: string
            enum: [en_attente, traite]
      responses:
        "200":
          description: Liste refus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Refus"
  /api/refus/{id}/traiter:
    put:
      operationId: traiterRefus
      tags: [refus]
      summary: Traiter un lot refoulé (décision)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TraiterRefusInput"
      responses:
        "200":
          description: Refus traité
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refus"
...
components:
  schemas:
...
    Refus:
      type: object
      required: [refus]
      properties:
        refus:
          type: object
          required: [id, cooperativeId, venteExportateurId, poidsRefuleKg, motifRefus, statut, createdAt]
          properties:
            id:
              type: integer
            cooperativeId:
              type: integer
            venteExportateurId:
              type: integer
            poidsRefuleKg:
              type: string
            motifRefus:
              type: ["string", "null"]
            decision:
              type: ["string", "null"]
              enum: [retour_stock, declassement, autre_acheteur, perte, null]
            statut:
              type: string
              enum: [en_attente, traite]
            traitePar:
              type: ["integer", "null"]
            traiteLe:
              type: ["string", "null"]
              format: date-time
            createdAt:
              type: string
              format: date-time
        vente:
          type: ["object", "null"]
          properties:
            exportateurId:
              type: integer
            dateVente:
              type: string
            produit:
              type: ["string", "null"]
            nombreSacsRefoules:
              type: ["integer", "null"]
            numeroBonSortie:
              type: ["string", "null"]
    RefusStats:
      type: object
      properties:
        totalRefulesKg:
          type: string
        totalTraitesKg:
          type: string
        totalEnAttenteKg:
          type: string
        nbEnAttente:
          type: integer
        nbTraites:
          type: integer
    TraiterRefusInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [retour_stock, declassement, autre_acheteur, perte]
        entrepotRetourId:
          type: integer
        ancienGrade:
          type: string
        nouveauGrade:
          type: string
        nouvelExportateurId:
          type: integer
        prixUnitaireNouveauFcfa:
          type: integer
        motifPerte:
          type: string
        pvConstat:
          type: boolean
        tauxHumidite:
          type: number
...