# lib/api-zod/src/generated/types/refus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoopDigital — API de gestion des coopératives cacaoyères
 * OpenAPI spec version: 0.1.0
 */
import type { RefusRefus } from './refusRefus';
import type { RefusVente } from './refusVente';

export interface Refus {
  refus: RefusRefus;
  /** @nullable */
  vente?: RefusVente;
}