# lib/api-zod/src/generated/types/refusStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoopDigital — API de gestion des coopératives cacaoyères
 * OpenAPI spec version: 0.1.0
 */

export interface RefusStats {
  totalRefulesKg?: string;
  totalTraitesKg?: string;
  totalEnAttenteKg?: string;
  nbEnAttente?: number;
  nbTraites?: number;
}